Update InsertNodeManager.cpp Define a function to call routines to insert function.
This routine calls one of the insertion routines implemented earlier to insert a node at a specific position. Calls are based on certain conditions. For example, if the list is empty call InsertRootNode() routine. In addition to placing a node at the first node call InsertAtFirstNode() routine. Likewise, to insert anywhere between the first and last node call InsertNodeAtPosition(). Finally, if you want to insert to the back of the node call InsertAtBackNode(). Insertion is based on one 
of these conditions.