Update DeleteNodeManager.cpp
Created a routine named DeleleNode(). This routine calls one of the deletion routines implemented earlier to delete a node at a specific position. Calls are based on certain conditions. For example, delete a node at the first node called DeleteAtFirstNode() routine. Likewise, to delete anywhere between the first and last node call DeleteBetweenNode(). Finally, if you want to delete to the back of the node call DeleteBackNode(). Deletion is based on one of these conditions. To delete a node list cannot be empty.